requirements.txt
google-generativeai>=0.3.0
langchain>=0.1.0
langchain-google-genai>=0.0.5
selenium>=4.10.0
webdriver-manager>=4.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pytest>=7.0.0
requests>=2.28.0
beautifulsoup4>=4.12.0
tenacity>=8.2.0
rich>=13.4.0
chromadb>=0.4.0
tqdm>=4.65.0  